requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
distro==1.9.0
fastapi==0.115.8
gitdb==4.0.12
GitPython==3.1.44
groq==0.18.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langchain==0.3.20
langchain-core==0.3.59
langchain-groq==0.2.5
langchain-text-splitters==0.3.8
langgraph==0.3.5
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.66
langsmith==0.3.42
lxml==5.4.0
MarkupSafe==3.0.2
narwhals==1.39.0
numpy==2.2.5
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
pillow==11.2.1
protobuf==6.30.2
pyarrow==20.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.24.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.45.3
streamlit==1.45.0
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.0
zstandard==0.23.0
# Visualization and Analytics
plotly==5.22.0
kaleido==0.2.1  # For exporting plotly charts

# Enhanced UI Components
streamlit-option-menu==0.3.13
streamlit-extras==0.5.0
#streamlit-aggrid==1.0.5  # For advanced data grids
streamlit-elements==0.1.0  # For Material UI components

# Export and Document Generation
reportlab==4.2.5  # For PDF generation
openpyxl==3.1.5  # For Excel export
python-pptx==1.0.2  # For PowerPoint export

# Additional Utilities
humanize==4.11.0  # For human-readable formatting
python-multipart==0.0.20  # For file uploads
watchdog==6.0.0  # For file system monitoring

# Rich Terminal Output (for debugging)
rich==13.9.4

# Data Processing
tabulate==0.9.0  # For table formatting

# Code Quality Analysis (optional)
pylint==3.3.3
black==24.10.0
flake8==7.1.1

# Security Analysis (optional)
bandit==1.8.0

# Performance Monitoring
psutil==6.1.1

# Caching
diskcache==5.6.3
